% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ruul.R
\name{Ruul.closure}
\alias{Ruul.closure}
\title{Closure the matrix of relation}
\usage{
Ruul.closure(r)
}
\arguments{
\item{r}{the matrix of relation}
}
\description{
Closure the matrix of relation
}
\examples{
r <- matrix(c(1.00,0.69,0.68, 0.69,1.00,0.69, 0.68,0.69,1.00), 
     nrow = 3, ncol = 3, byrow = TRUE)
rt <- Ruul.closure(r)
rt
}
