% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ruul.R
\name{Ruul_dist}
\alias{Ruul_dist}
\title{Distance the matrix of data}
\usage{
Ruul_dist(xstd, dist = 1)
}
\arguments{
\item{xstd}{the standard matrix of data}

\item{dist}{the option to distance the matrix of data \cr
- case 1: scalar product similarity \cr
- case 2: cosine similarity \cr 
- case 3: correlation coefficient similarity \cr
- case 4: correlation exponent coefficient similarity \cr
- case 5: Max-Min similarity \cr
- case 6: arithmetic average minimum similarity \cr
- case 7: geometry average minimum similarity \cr
- case 8: Euclidean distance similarity \cr
- case 9: Hamming distance similarity \cr
- case 10: Chebyshev distance similarity \cr
- case 11: Euclidean reciprocal distance similarity \cr
- case 12: Hamming reciprocal distance similarity \cr
- case 13: Chebyshev reciprocal distance similarity \cr
- case 14: Euclidean exponent distance similarity \cr
- case 15: Hamming exponent distance similarity \cr
- case 16: Chebyshev exponent distance similarity \cr}
}
\description{
Distance the matrix of data
}
\examples{
xstd <- matrix(c(1.00,0.81,0.79, 0.81,1.00,0.82, 0.79,0.82,1.00), 
     nrow = 3, ncol = 3, byrow = TRUE)
r <- Ruul_dist(xstd, dist = 1)
r
}
