% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ruul.R
\name{Ruul_scale}
\alias{Ruul_scale}
\title{Scale the matrix of data}
\usage{
Ruul_scale(x, scale = 1)
}
\arguments{
\item{x}{the matrix of data}

\item{scale}{the option to scale the matrix of data \cr
- case 0: unchanged \cr
- case 1: zero-mean normalization \cr
- case 2: Min-Max normalization \cr
- case 3: non-liner normalization \cr}
}
\description{
Scale the matrix of data
}
\examples{
x <- matrix(c(1.24,1.72, 1.36,1.74, 1.38,1.64), 
     nrow = 3, ncol = 2, byrow = TRUE)
xstd <- Ruul_scale(x, scale = 1)
xstd
}
